Add S₀ literals
A literal is a compile-time binary constant.  It will be used to create
constants of various types; each type will have a constructor that knows
how to parse a binary literal into an instance of the type (preferrably
via a macro, so that we can catch syntax errors at compile time).