Add S₀ methods
A method is similar to a closure.  It only contains one block, though
(so it's not a linear additive sum), and it must be part of an object to
be invoked.  The object itself is passed in as one of the input
parameters, giving the method control over whether it's passed on to the
return continuation or freed.