Add S₀ invocations
An invocation is the final "command" in a basic block; it transfers
control to some other piece of code, which is either a closure or the
method of some object.  The state of the current environment just before
the invocation must satisfy the input requirements of the closure or
method, without any additional elements in the environment.