Add initial YAML S₀ parser
There's no proper documentation of the format yet, but it follows in the
obvious way from the data type declarations.  (This is only temporary,
of course!  We need to document the format ASAP.)

This patch includes a test harness, which reads a bunch of expected-good
and expected-bad parses from some YAML files.  That way, the test suite
can be easily reused across whatever new hosts we develop next.  The
test suite is currently incomplete, since there are some checks that we
don't yet perform, but that we'll need to.  (That means that this first
C host is a special case, since we'll keep those checks out of the test
suite until they're implemented.  Later hosts will have to use the test
suite as-is, so they'll need their own way to keep track of which of the
test cases they're temporarily failure while they're first being
developed.)