[c] Create environment types from input mappings
Now that an input mapping contains a type for each entry, we can extract
two environment types from a mapping: an *external* one and an
*internal* one.  The external one (which contains the `from` names from
the mapping) represents what the caller must provide when invoking a
block.  The internal one (which contains the `to` names from the
mapping) represents what the block will see when it first starts
executing.