[c] Include type in name mappings
The name mapping, which defines the inputs to a block (and optionally
how they are renamed), now also includes an entity type for each entry.
Right now, the only possible entity type is `any`, which has the same
effect as not having the types in the name mappings at all.  But
eventually, we'll have more specific types that you can use to restrict
what values can be passed in to the block.