[c] Verify environment when parsing statements
We now keep track of the current environment type of a block while
parsing the block's statement list.  This lets us verify that the
block's inputs, and each of its statements, are well-formed.  For
statements, that really just means that we ensure that nothing gets
overwritten in the environment.  The story will be more complex for
invocations, since we'll also need to verify that the target is
something that we can call, and that we provide the target with all of
the inputs that it expects.