[c] Update an environment type with an invocation
This new function ensures that an environment type satisfies the
prerequisites of an invocation, and then updates the environment type
based on what that invocation would do.  Right now we only have `any`
types for environment entries, so we can only verify that the requested
target exists in the environment.  Once we have more detailed types
we'll be able to perform more precise checks, too.