[c] Check subtype relationship ⊆
This patch adds two functions for checking whether one type is a subtype
of another — one function for entity types, and the other for
environment types.  We'll typically call the operands `have` and
`requires`, because of how we'll use this relationship during
type-checking: if the `have` type is a subtype of the `requires` type,
then if the caller can guarantee that they'll provide a value from the
`have` type, then we can assume that the value will also satisfy the
`requires` type.

Note that we're stricter than you might expect when checking environment
types; the set of names in both environment types must be exactly the
same.