[c] Add closure entity types
The next entity type (and the first actually useful one!) is a closure
type.  You have to specify an environment type for each of the branches
that you expect the closure to have.  An entity satisfies this type if
it's a closure, has exactly the set of branches that you expect, and
each branch's input mapping satisfies the corresponding environment type
in the closure type.