Move renaming to invocations
Before, you could apply a "renaming" to the inputs of a block, so that
the caller of the block could provide one set of names, but the inside
of the block would those values at a different set of names.  We've
moved things around so that this renaming is now defined when you
*invoke* the block (via `invoke-closure` or `invoke-method`), instead of
when you *define* the block.