[c] Remove `self_input` from methods
Before, a method declared a `self_input` in addition to the normal set
of inputs defined in the method's block.  The intention was that when
invoking a method, we'd pass in the object containing the method as an
additional entry in the environment, with `self_input` telling us which
name to place it at.  We don't need to treat this specially, though; you
can (er, must) now just include the "self" input along with any others
in the method's block's `inputs` section.  Once we start type-checking
method invocations, we'll make sure that there's an entry in the
`parameters` section mapping the containing object to one of the inputs.