[c] Add object entity type
Object types look a lot like environment types, since they map names to
entity types.  They're interesting because they're not as restrictive.
For environment types, you must have *exactly* the same set of names in
the environment as are in the time — no more, no less.  For objects,
you're allowed to have more stuff in the object than the object type
requires, since whenever you free the object itself, all of the stuff
inside of it is freed, too, whether or not you knew it existed!