[c] Verify type of `invoke-method` target
We can now verify the types of an `invoke-method` invocation, like we
can with `invoke-closure`.  The target must be an object containing a
method with the desired name.  Unlike `invoke-closure`, the target is
*not* removed from the environment before passing control to it;
instead, one of the method's inputs must be a "self" parameter, which
will be bound to the object containing the method.