[c] Rename entries in an environment
You can now apply a renaming (giving by a name mapping) to an
environment, just like you could already do to an environment type.  We
need this when interpreting an invocation, to translate the parameters
of the closure or method so that they use the names expected inside its
body.