[c] Add an initial S₀ interpreter
We can finally execute some S₀ code!  We have to introduce the notion of
a *primitive method*, which is just like a regular method, but which is
implemented in the host, instead of by a block of S₀ code.  This is how
we provide the "escape hatch" `finish` method, which S₀ can call to
return control back to the host.