[c] Add extractor closure
This adds a new primitive "extractor" closure.  This finishes the
current executor, like the existing primitive `finish` method, but also
requires the S₀ to pass in an entity as a parameter to the closure.
This entity is "extracted" from the S₀ environment and passed back to
the host code that started the execution.